requirements.txt
Flask
Redis
gevent
requests
Pillow
numpy
